requirements.txt
imageio>=2.6.1
torch>=1.5.1
matplotlib>=3.3.3
torchvision>=0.6.1
scikit_image
numpy>=1.19.5
opencv_contrib_python>=4.1.2.30
pillow
